refactor: separate code into classes and build them into one userscript file
