+++ build.js
@@ -1,0 +1,8 @@
+import { build } from "@kellnerd/userscript-bundler";
+
+build({
+	userscriptSourcePath: "src/",
+	// docSourcePath: "doc/",
+	outputPath: "dist/",
+	readmePath: "doc/discard.md",
+});
