+++ src/voidverified.meta.js
@@ -1,0 +1,9 @@
+/** @type {import('@kellnerd/userscript-bundler').EnhancedUserscriptMetadata} */
+const metadata = {
+	name: "VoidVerified",
+	author: "voidnyan", // optional, falls back to your GitHub username
+	description: "Display a verified sign next to user's name in AniList.",
+	match: "https://anilist.co/*",
+};
+
+export default metadata;
